+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+- `IntId` is now a trait and not an abstract class.
 - Minimum requirement Doctrine ORM 3.2.
 
 ## [7.1.2] - 2024-02-20